app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
Flask-WTF==1.2.1
markdown==3.7
bleach==6.2.0
cryptography==45.0.6
